+++ lbs/main.c
@@ -43,11 +43,11 @@ main(int argc, char * argv[])
 	/* Command-line parameters. */
 	char * opt_s = NULL;
 	char * opt_d = NULL;
-	intmax_t opt_b = -1;
-	intmax_t opt_n = 16;
+	size_t opt_b = (size_t)(-1);
+	size_t opt_n = 16;
 	char * opt_p = NULL;
 	int opt_1 = 0;
-	intmax_t opt_l = 0;
+	long opt_l = 0;
 	int opt_L = 0;
 
 	/* Working variables. */
@@ -60,7 +60,7 @@ main(int argc, char * argv[])
 	while ((ch = GETOPT(argc, argv)) != NULL) {
 		GETOPT_SWITCH(ch) {
 		GETOPT_OPTARG("-b"):
-			if (opt_b != -1)
+			if (opt_b != (size_t)(-1))
 				usage();
 			if (PARSENUM(&opt_b, optarg, 512, 128 * 1024)) {
 				warn0("Block size must be in [2^9, 2^17]");
@@ -135,7 +135,7 @@ main(int argc, char * argv[])
 		usage();
 	if (opt_d == NULL)
 		usage();
-	if (opt_b == -1)
+	if (opt_b == (size_t)(-1))
 		usage();
 
 	/* Resolve the listening address. */
